+++ Pipfile
@@ -1,12 +1,20 @@
 [[source]]
+name = "pypi"
 url = "https://pypi.org/simple"
 verify_ssl = true
-name = "pypi"
 
 [packages]
+fastapi = "*"
 requests = "*"
+uvicorn = "*"
 
 [dev-packages]
+flake8 = "*"
+gunicorn = "*"
 
 [requires]
 python_version = "3.9"
+
+[scripts]
+dev = "uvicorn main:app --workers 4 --port 3000 --reload"
+start = "gunicorn --bind 0.0.0.0:3000 -w 4 -k uvicorn.workers.UvicornWorker main:app"
